+++ README.md
@@ -1,5 +1,6 @@
-# No-code Test Platform
+# No-code QA Platform
 
 ## Tech stack:
 - Java 17
 - Spring Boot 3
+- RocksDB 8.5.3
