+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'no-code-test-platform'
+rootProject.name = 'no-code-qa-platform'
