+++ src/main/java/apps/amaralus/qa/platform/rocksdb/RocksDbRuntimeException.java
@@ -1,4 +1,4 @@
-package org.example.nocodetestplatform.rocksdb;
+package apps.amaralus.qa.platform.rocksdb;
 
 public class RocksDbRuntimeException extends RuntimeException {
     public RocksDbRuntimeException(Throwable cause) {
