+++ src/main/java/apps/amaralus/qa/platform/rocksdb/config/RocksDbInitializationException.java
@@ -1,4 +1,4 @@
-package org.example.nocodetestplatform.rocksdb.config;
+package apps.amaralus.qa.platform.rocksdb.config;
 
 public class RocksDbInitializationException extends RuntimeException {
     public RocksDbInitializationException(Throwable cause) {
